+++ Repositories/ProjectRepository.cs
@@ -32,6 +32,7 @@ using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Net;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
@@ -121,6 +122,15 @@ namespace Repositories
         void CreateProjectIndex();
         void DeleteIndex();
         void MigrateDatabase(List<Project> projectsToExport);
+
+        /// <summary>
+        ///     This method will call the ElasticSearch index to retrieve projects where the title starts with the query.
+        /// </summary>
+        /// <param name="query">The string of characters with which the title must begin</param>
+        /// <returns>
+        ///     This method return a list of projects where the title, or part of the title matches the query.
+        /// </returns>
+        Task<List<Project>> FindProjectsWhereTitleStartsWithQuery(string query);
     }
 
     /// <summary>
@@ -708,6 +718,30 @@ namespace Repositories
             taskPublisher.RegisterTask(Newtonsoft.Json.JsonConvert.SerializeObject(eSProjectDTO), Subject.ELASTIC_CREATE_OR_UPDATE);
 
 
+        }
+
+        /// <summary>
+        ///     This method will call the ElasticSearch index to retrieve projects where the title starts with the query.
+        /// </summary>
+        /// <param name="query">The string of characters with which the title must begin</param>
+        /// <returns>
+        ///     This method return a list of projects where the title, or part of the title matches the query.
+        /// </returns>
+        public async Task<List<Project>> FindProjectsWhereTitleStartsWithQuery(string query)
+        {
+            RestRequest request = new RestRequest("_search", Method.POST);
+            string body = queries.SearchProjectsByPartOfTitle
+                .Replace("ReplaceWithQuery", query);
+            request.AddParameter("application/json", body, ParameterType.RequestBody);
+            IRestResponse restResponse = elasticSearchContext.Execute(request);
+            if(!restResponse.IsSuccessful && restResponse.Content == "")
+            {
+                throw new ElasticUnavailableException();
+            }
+            List<ESProjectDTO> esProjectFormats = new List<ESProjectDTO>();
+
+            ParseJsonToESProjectFormatDTOList(restResponse, esProjectFormats);
+            return await ConvertProjects(esProjectFormats);
         }
         public Task<bool> ProjectExistsAsync(int id)
         {