+++ Services/Services/ProjectService.cs
@@ -84,6 +84,12 @@ namespace Services.Services
         /// <returns>The project with users and collaborators</returns>
         Task<List<Project>> GetAllWithUsersCollaboratorsAndInstitutionsAsync();
 
+        /// <summary>
+        ///     Get projects where the title begins with a certain string of characters.
+        /// </summary>
+        /// <param name="Query">The string with which the title must begin</param>
+        /// <returns>The projects where the title matches the query</returns>
+        Task<List<Project>> FindProjectsWhereTitleStartsWithQuery(string query);
     }
 
     /// <summary>
@@ -247,6 +253,10 @@ namespace Services.Services
         public Task<IEnumerable<Project>> GetUserProjects(int userId)
         {
             return Repository.GetUserProjects(userId);
+        }
+        public async Task<List<Project>> FindProjectsWhereTitleStartsWithQuery(string query)
+        {
+            return await Repository.FindProjectsWhereTitleStartsWithQuery(query);
         }
     }